# store/cachekv, x/bank/types: algorithmically fix pathologically slow code