@odeke-em 2021-02-27
Thank you for the reviews and kind words, @alessio @fedekunze @zmanian.

 >Mind also adding an improvement changelog entry?

Done, @fedekunze, please take a look and thank you.